+++ R/reshape_chroms.R
@@ -1,19 +1,19 @@
-
 #' Reshapes list of chromatograms from wide to long format
 #' @name reshape_chroms
 #' @param x A list of chromatographic matrices in wide format.
 #' @param idx Indices of chromatograms to convert
 #' @param sample_var String with name of new column containing sample IDs.
+#' @param lambdas Wavelength(s) to include.
 #' @return A list of chromatographic matrices in long format.
 #' @author Ethan Bass
 #' @export
 
-reshape_chroms <- function(x, idx, sample_var = "sample"){
+reshape_chroms <- function(x, idx, sample_var = "sample", lambdas){
   if (missing(idx)){
     idx <- seq_along(x)
   }
   dat <- lapply(idx, function(i){
-    xx <- reshape_chrom(x[[i]])
+    xx <- reshape_chrom(x[[i]], lambdas)
     xx[,sample_var] <- names(x)[[i]]
     xx
   })
@@ -24,13 +24,17 @@ reshape_chroms <- function(x, idx, sample_var = "sample"){
 #' @name reshape_chrom
 #' @importFrom stats reshape
 #' @param x A chromatographic matrix in wide format.
+#' @param lambdas Wavelength(s) to include.
 #' @return A chromatographic matrix in long format.
 #' @author Ethan Bass
 #' @noRd
-reshape_chrom <- function(x){
+reshape_chrom <- function(x, lambdas){
   if (ncol(x) == 1)
     stop("The provided data is already in long format!")
   x <- as.data.frame(x)
+  if (!missing(lambdas)){
+    x <- x[,lambdas,drop=FALSE]
+  }
   data <- reshape(as.data.frame(rt=rownames(x),x), direction = "long",
                   varying = list(1:ncol(x)), v.names="absorbance",
                   times = colnames(x), timevar = "lambda",
@@ -41,19 +45,27 @@ reshape_chrom <- function(x){
   data[,c(3,2,1)]
 }
 
-# 
-# x<-pk_tab
-# reshape_peaktable <- function(x){
-#   x$tab <- reshape(as.data.frame(chr=rownames(x$tab), x$tab), direction = "long",
-#     varying = list(1:ncol(x$tab)), v.names = x$args[["response"]],
-#     times = colnames(x$tab), timevar = "peak",
-#     idvar = "chr", ids = rownames(x$tab))
-#   rownames(x$tab) <- NULL
-#   x$tab <- x$tab[,c(3,1,2)]
-#   merge(x$tab, x$sample_meta, )
-#   x
-# }
-# 
-# x$tab %>% filter(peak=="V1") %>% ggplot(aes(x=chr, y=area)) + geom_violin() +
-#   geom_dotplot(binaxis='y', stackdir='center', stack.ratio=1.5)
+#' Reshapes peak table from wide to long format
+#' @name reshape_peaktable
+#' @param x A \code{peak_table} object.
+#' @param peaks A character vector specifying peaks to include.
+#' @return A data.frame containing the information for the specified peaks in
+#' long format.
+#' @author Ethan Bass
+#' @export
 
+reshape_peaktable <- function(x, peaks){
+  if (!missing(peaks)){
+    x$tab <- x$tab[,which(colnames(x$tab) %in% peaks), drop=FALSE]
+  }
+  xx <- reshape(as.data.frame(chr=rownames(x$tab), x$tab), direction = "long",
+    varying = list(1:ncol(x$tab)), v.names = x$args[["response"]],
+    times = colnames(x$tab), timevar = "peak",
+    idvar = "sample", ids = rownames(x$tab))
+  rownames(xx) <- NULL
+  xx <- xx[,c(3,1,2)]
+  xx <- merge(xx, data.frame(sample=row.names(x$tab), x$sample_meta), by="sample", all.x = TRUE)
+  xx
+}
+
+